requirements.txt
numpy
Flask
psycopg2-binary
requests
wget
gdal

